requirements.txt
pandas==2.3.2
requests==2.32.5
igdb-api-v4==0.3.3
boto3==1.40.48
awswrangler==3.13.0